Move terser option into crate
